fix: introduce checks for ca and client certs
