+++ packages/insomnia/src/common/validators.ts
@@ -1,11 +1,18 @@
+import type { CaCertificate } from "../models/ca-certificate";
+import type { ClientCertificate } from "../models/client-certificate";
 import type { Settings } from "../models/settings";
 import type { RenderedRequest } from "../templating/types";
 
-export function isFsAccessingAllowed(renderedRequest: RenderedRequest, settings: Settings) {
+export function isFsAccessingAllowed(
+  renderedRequest: RenderedRequest,
+  settings: Settings,
+  clientCertificates: ClientCertificate[],
+  caCert?: CaCertificate | null,
+) {
   const throwError = (fileName: string) => {
     throw `Insomnia cannot access the file ‘${fileName}’. You can adjust this in Preferences → Security.`;
   }
-  
+
   // case1: check request body (set by scripts or request body editor)
   if (renderedRequest.body.fileName !== undefined && renderedRequest.body.fileName !== '') {
     const allowed = settings?.dataFolders.some(folder => folder !== '' && renderedRequest.body.fileName?.startsWith(folder));
@@ -26,5 +33,31 @@ export function isFsAccessingAllowed(renderedRequest: RenderedRequest, settings:
     });
   }
 
-  // case3: check "file" template tags, which is checked in tag implementation
+  // case3: check the caCert path
+  if (!caCert?.disabled && caCert?.path) {
+    const allowed = settings?.dataFolders.some(folder => folder !== '' && caCert.path?.startsWith(folder));
+    if (!allowed) {
+      throwError(caCert.path);
+    }
+  }
+
+  // case4: check paths of client certificates
+  if (Array.isArray(clientCertificates)) {
+    clientCertificates.forEach(cert => {
+      if (cert.disabled) {
+        return;
+      }
+
+      [cert.key, cert.cert, cert.pfx].forEach(targetPath => {
+        if (targetPath) {
+          const allowed = settings?.dataFolders.some(folder => folder !== '' && targetPath !== "" && targetPath?.startsWith(folder));
+          if (!allowed) {
+            throwError(targetPath);
+          }
+        }
+      });
+    });
+  }
+
+  // case5: check "file" template tags, which is checked in tag implementation
 }
